uploaded .aup & .mp3 files without data folder
I suspect the .aup. file can't be read properly without its accompanying data folder. To amend this, I aligned the separate sample tracks end to end in an .mp3 file so that it can be listened to. And only now do I recall that I can zip the data files together. But I'll try that on another branch and merge, because I'm tired and my bed yells at me often to be settled. And my bear gets lonely.